+++ common/src/main/java/immersive_melodies/client/sound/SoundManager.java
@@ -10,4 +10,6 @@ public interface SoundManager {
     boolean isFirstPerson(Entity entity);
 
     boolean audible(Entity entity);
+
+    void pauseGameMusic();
 }