+++ common/src/main/java/immersive_melodies/mixin/MusicTrackerAccessor.java
@@ -1,0 +1,12 @@
+package immersive_melodies.mixin;
+
+import net.minecraft.client.sound.MusicTracker;
+import net.minecraft.client.sound.SoundInstance;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.gen.Accessor;
+
+@Mixin(MusicTracker.class)
+public interface MusicTrackerAccessor {
+    @Accessor("current")
+    SoundInstance getCurrent();
+}